Added tmp solution to deps issue with windows

+++ src/main.odin
@@ -5,6 +5,7 @@ import "core:mem"
 
 DEBUG_REMOVE_CONFIG := false
 DEBUG_TRACK_ALLOC := false
+DEBUG_NO_DEPS := false
 
 main :: proc() {
 
@@ -35,15 +36,17 @@ main :: proc() {
 		mem.tracking_allocator_destroy(&track)
 	}
 
-	missing_dependancies := Check_Dependancies({"yt-dlp", "eyeD3"})
-	defer {
-		for m in missing_dependancies do delete_string(m)
-		delete_slice(missing_dependancies)
-	}
-	if len(missing_dependancies) > 0 {
-		fmt.eprintln("[ERROR] Failed Dependancy Check, could not find:")
-		for m in missing_dependancies do fmt.eprintfln("\t-  %s", m)
-		return
+	if !DEBUG_NO_DEPS {
+		missing_dependancies := Check_Dependancies({"yt-dlp", "eyeD3"})
+		defer {
+			for m in missing_dependancies do delete_string(m)
+			delete_slice(missing_dependancies)
+		}
+		if len(missing_dependancies) > 0 {
+			fmt.eprintln("[ERROR] Failed Dependancy Check, could not find:")
+			for m in missing_dependancies do fmt.eprintfln("\t-  %s", m)
+			return
+		}
 	}
 
 	Globals_Init()
